requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
fastapi==0.115.8
filelock==3.18.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.9
idna==3.10
invoke==2.2.0
Jinja2==3.1.5
jiter==0.9.0
loguru==0.7.3
MarkupSafe==3.0.2
multidict==6.2.0
nodeenv==1.9.1
openai==1.64.0
platformdirs==4.3.7
pre_commit==4.1.0
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
ruff==0.9.6
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
yarl==1.18.3
